Split MCP file import tests
